grocery list screen created; realm added;

+++ Fooder/IngridientObject.swift
@@ -9,13 +9,13 @@
 import Foundation
 
 class Ingridient{
-    let id: Double
+    let id: Int
     let amount: Double
     let unit: String
     let unitShort: String
     let name: String
     
-    init(id: Double, amount: Double, unit: String, unitShort: String, name: String) {
+    init(id: Int, amount: Double, unit: String, unitShort: String, name: String) {
         self.id = id
         self.amount = amount
         self.unit = unit
